Remove onFrame function
