+++ apps/typegpu-docs/src/content/examples/algorithms/mnist-inference/index.ts
@@ -253,7 +253,13 @@ function resetDrawing() {
   }
 }
 
+let disposed = false;
+
 function run() {
+  if (disposed) {
+    return;
+  }
+
   const scale = canvas.width / SIZE;
 
   context.clearRect(0, 0, canvas.width, canvas.height);
@@ -410,6 +416,7 @@ export const controls = {
 // #region Resource cleanup
 
 export function onCleanup() {
+  disposed = true;
   root.destroy();
   root.device.destroy();
 }
