+++ apps/typegpu-docs/src/content/examples/image-processing/blur/index.ts
