+++ apps/typegpu-docs/src/content/examples/image-processing/camera-thresholding/index.ts
@@ -1,7 +1,3 @@
-// -- Hooks into the example environment
-import { onFrame } from '@typegpu/example-toolkit';
-// --
-
 import { f32, vec2f } from 'typegpu/data';
 import tgpu, { asUniform, builtin, wgsl } from 'typegpu/experimental';
 
@@ -103,28 +99,38 @@ export const controls = {
   },
 };
 
-onFrame(() => {
-  if (!(video.currentTime > 0)) {
+let disposed = false;
+
+function run() {
+  if (disposed) {
     return;
   }
 
-  renderProgram.execute({
-    colorAttachments: [
-      {
-        view: context.getCurrentTexture().createView(),
-        clearValue: [0, 0, 0, 1],
-        loadOp: 'clear',
-        storeOp: 'store',
-      },
-    ],
+  if (video.currentTime > 0) {
+    renderProgram.execute({
+      colorAttachments: [
+        {
+          view: context.getCurrentTexture().createView(),
+          clearValue: [0, 0, 0, 1],
+          loadOp: 'clear',
+          storeOp: 'store',
+        },
+      ],
 
-    vertexCount: 6,
-  });
+      vertexCount: 6,
+    });
 
-  root.flush();
-});
+    root.flush();
+  }
+
+  requestAnimationFrame(run);
+}
+
+run();
 
 export function onCleanup() {
+  disposed = true;
+
   if (video.srcObject) {
     for (const track of (video.srcObject as MediaStream).getTracks()) {
       track.stop();
