+++ apps/typegpu-docs/src/content/examples/rendering/box-raytracing/index.ts
@@ -1,5 +1,5 @@
 // -- Hooks into the example environment
-import { addSliderPlumParameter, onFrame } from '@typegpu/example-toolkit';
+import { addSliderPlumParameter } from '@typegpu/example-toolkit';
 // --
 
 import { arrayOf, bool, f32, struct, u32, vec3f, vec4f } from 'typegpu/data';
@@ -295,6 +295,23 @@ const renderPipeline = root.makeRenderPipeline({
   },
 });
 
+let disposed = false;
+
+const onFrame = (loop: (deltaTime: number) => unknown) => {
+  let lastTime = Date.now();
+  const runner = () => {
+    if (disposed) {
+      return;
+    }
+    const now = Date.now();
+    const dt = now - lastTime;
+    lastTime = now;
+    loop(dt);
+    requestAnimationFrame(runner);
+  };
+  requestAnimationFrame(runner);
+};
+
 onFrame((deltaTime) => {
   root.setPlum(canvasWidthPlum, canvas.width);
   root.setPlum(canvasHeightPlum, canvas.height);
@@ -321,6 +338,7 @@ onFrame((deltaTime) => {
 });
 
 export function onCleanup() {
+  disposed = true;
   root.destroy();
   root.device.destroy();
 }
